Re-added citadel reinforcement statistics
Changes to Namelayer plugin, Citadel Changes also required.

These changes allow for groups to keep track of how many reinforcements
of each type are under that group. The groups will now keep track of the
number, and now only incriment or decriment from that number when a
reinforcement is added or removed from the DB.